great_reports/requirements.txt
Django==1.11.8
djangorestframework==3.8.2
django-rest-knox==3.1.4
django-webpack-loader==0.6.0
psycopg2==2.7.3.2